Migrate to emergent goal hierarchies in curriculum training
Replace explicit Goal objects with example-driven PFC training.

Changes:
- DELETE: All explicit Goal creation methods (~287 lines)
  - _setup_stage_goal_hierarchies() and _setup_stage3/4_goals()
  - All _create_*_goal() methods (tower_hanoi, essay, maze, etc.)
  - get_goal_hierarchy() and set_active_goal_hierarchy()
  - _goal_hierarchies Dict[str, Goal] cache

- ADD: Emergent goal training infrastructure (~380 lines)
  - _train_emergent_goal_hierarchies() - Main coordinator
  - _train_stage3/4_emergent_goals() - Stage-specific training
  - 6 training example generators creating neural patterns
  - _train_pfc_transitions() - Hebbian learning loop

Training Approach:
- Abstract (rostral) + concrete (caudal) PFC pattern examples
- Hebbian learning: abstractconcrete transitions
- Dopamine reinforcement for valuable patterns
- Goals emerge from learned WM dynamics (no symbolic objects)

Hierarchies Trained:
Stage 3: Tower of Hanoi, Essay writing, Maze solving
Stage 4: Raven's matrices, Hypothesis testing, Multi-premise reasoning

Benefits:
- Biologically plausible (local learning rules)
- Flexible (learned from data, not programmed)
- Consistent with RL migration philosophy
- Reduces code complexity (~650 lines deleted total)

FIX: Pre-existing mypy errors
- Add source_inputs field to LayeredCortexState
- Fix None assignments with type: ignore annotations
- Fix forward() signature override in multisensory
- Fix return type in get_goal_value()
- Remove invalid dt_ms parameter in working_memory
- Add type ignores for value_estimates attribute

Related: Completes Phase 2 of goal hierarchy emergence migration

+++ src/thalia/regions/prefrontal/prefrontal.py
@@ -1334,7 +1334,7 @@ class Prefrontal(NeuralRegion):
             raise ConfigurationError("Emergent goals not enabled.")
 
         # Compute value as dot product with learned value weights
-        return torch.sum(goal_pattern * self.emergent_goals.value_weights).item()
+        return float(torch.sum(goal_pattern * self.emergent_goals.value_weights).item())
 
     def reset_goal_tags(self) -> None:
         """Reset synaptic tags for goal patterns.
